pyproject.toml
[tool.poetry]
name = "ailearn"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.16.2"
requests = "^2.31.0"
Django = "^5.0.4"
channels = "^4.1.0"
daphne = "^4.1.0"
python-dotenv = "^1.0.1"
fontawesomefree = "^6.5.1"
pillow = "^10.3.0"
markdown = "^3.6"
pymdown-extensions = "^10.7.1"
pyfunctional = "^1.5.0"
numpy = "1.26.4"
wordcloud = "^1.9.3"
mysqlclient = "^2.2.4"
crypto = "^1.4.1"
wheel = "^0.43.0"
jwcrypto = "^1.5.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
